Revert "Rewrite MSTest unit test project to nUnit (#8)"
This reverts commit bd95091525b81c3d77340557e74faa92544ec91c.
