+++ .gitignore
@@ -153,9 +153,6 @@ coverage*.info
 *.coverage
 *.coveragexml
 
-## Code Coverage report
-coverage/
-
 # NCrunch
 _NCrunch_*
 .*crunch*.local.xml
