+++ NemzetiVirusbolt.ValidationTests/NemzetiVirusbolt.ValidationTests.csproj
@@ -9,9 +9,9 @@
   <ItemGroup>
     <PackageReference Include="FluentValidation.Validators.UnitTestExtension" Version="1.9.0.13" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
-    <PackageReference Include="coverlet.collector" Version="3.0.2" />
-    <PackageReference Include="NUnit" Version="3.13.2" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
+    <PackageReference Include="MSTest.TestAdapter" Version="2.1.1" />
+    <PackageReference Include="MSTest.TestFramework" Version="2.1.1" />
+    <PackageReference Include="coverlet.collector" Version="1.3.0" />
   </ItemGroup>
 
   <ItemGroup>
