--- scripts/run-unit-test.sh
@@ -1,13 +1,0 @@
-﻿#!/usr/bin/bash
-
-cd ../NemzetiVirusbolt.Desktop.Tests
-
-rm -rf ./coverage
-
-dotnet test
-
-dotnet tool install -g dotnet-reportgenerator-globaltool
-
-dotnet test --collect:"XPlat Code Coverage"
-
-reportgenerator "-reports:./**/coverage.cobertura.xml" "-targetdir:./coverage" "-reporttypes:Html"
